requirements.txt
fastapi==0.111.1
uvicorn[standard]==0.30.1
python-multipart==0.0.9
websockets==12.0
starlette==0.37.2
slowapi==0.1.9
supabase==2.4.3
gotrue==2.4.3
postgrest==0.16.7
storage3==0.7.4
realtime==1.0.6
httpx==0.27.0
PyJWT==2.8.0
bcrypt==4.1.3
pyotp==2.9.0
pycryptodome==3.20.0
cryptography==42.0.8
weasyprint==62.3
qrcode[pil]==7.4.2
Pillow==10.3.0
Jinja2==3.1.4
opentimestamps-client==0.7.1
pySMART==1.3.0
psutil==5.9.8
numpy==1.26.4
scipy==1.13.1
python-dotenv==1.0.1
pydantic==2.7.1
python-dateutil==2.9.0
pytz==2024.1
apscheduler==3.10.4
click==8.1.7
pytest==8.2.1
pytest-asyncio==0.23.7
pytest-cov==5.0.0
httpx==0.27.0
ruff==0.4.7